move client pkg to pkg/daemon/client
