+++ cmd/build.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"errors"
 
-	"github.com/ipfs/testground/client"
 	"github.com/ipfs/testground/pkg/api"
+	"github.com/ipfs/testground/pkg/daemon/client"
 	"github.com/ipfs/testground/pkg/engine"
 	"github.com/ipfs/testground/pkg/util"
