+++ pkg/daemon/doc.go
@@ -1,0 +1,1 @@
+package daemon
