+++ pkg/server/describe.go
@@ -6,8 +6,8 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/ipfs/testground/client"
 	"github.com/ipfs/testground/pkg/api"
+	"github.com/ipfs/testground/pkg/daemon/client"
 	"github.com/ipfs/testground/pkg/tgwriter"
 	"go.uber.org/zap"
 )
